Added food page to layout
