+++ small-talk/pages/food/food.test.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import Food from './page'; // Ensure the correct file path and component name are used
-import ThemeLayout from '../components/ThemeLayout'; // Ensure correct path
-import MenuTabs from '../components/foodPage/MenuTabs'; // Ensure correct path
+import ThemeLayout from '../../app/components/ThemeLayout'; // Ensure correct path
+import MenuTabs from '../../app/components/foodPage/MenuTabs'; // Ensure correct path
 
 jest.mock('../components/foodPage/MenuTabs', () => {
     const MenuTabsMock = () => <div>MenuTabs Mock</div>;
